# Load trabajos by default in Auto to improve performance